+++ pilot/pkg/bootstrap/server.go
@@ -1114,7 +1114,10 @@ func (s *Server) initMulticluster(args *PilotArgs) {
 	if s.kubeClient == nil {
 		return
 	}
-	s.multiclusterController = multicluster.NewController(s.kubeClient, args.Namespace, s.clusterID, s.environment.Watcher)
+	s.multiclusterController = multicluster.NewController(s.kubeClient, args.Namespace, s.clusterID, s.environment.Watcher, func(r *rest.Config) {
+		r.QPS = args.RegistryOptions.KubeOptions.KubernetesAPIQPS
+		r.Burst = args.RegistryOptions.KubeOptions.KubernetesAPIBurst
+	})
 	s.XDSServer.ListRemoteClusters = s.multiclusterController.ListRemoteClusters
 	s.addStartFunc("multicluster controller", func(stop <-chan struct{}) error {
 		return s.multiclusterController.Run(stop)
